backend/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
django-resized==1.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.3.0
protobuf==4.21.7
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
